Fix bug: undesired escape on double quote
When a string is not closed, which expands to end of buffer, I'd
terminate it by inserting a double quote, but `lisp-electric-insert-
double-quote` make it escaped.

To fix this, it checks if the string expands to end of buffer and
if so it doesn't escape the double quote.

This solution doesn't work if there's a double quote formally in a
comment which happen to terminate the non-terminated string.